+++ src/ophyd_async/tango/demo/__init__.py
@@ -4,6 +4,7 @@ from ._motor import DemoMotor
 from ._point_detector import DemoPointDetector
 from ._point_detector_channel import DemoPointDetectorChannel, EnergyMode
 from ._stage import DemoStage
+from ._tango_device_servers import predict_trl, tango_device_servers_args
 
 __all__ = [
     "DemoMotor",
@@ -11,4 +12,6 @@ __all__ = [
     "EnergyMode",
     "DemoPointDetectorChannel",
     "DemoPointDetector",
+    "predict_trl",
+    "tango_device_servers_args",
 ]
