+++ src/ophyd_async/tango/testing/_tango.py
@@ -1,4 +1,4 @@
-"""Every Tango device server this repo ships for testing/demoing ophyd-async.
+"""Test-only Tango device servers this repo ships for testing ophyd-async.
 
 Kept as a single, self-contained module (not a package) so it stays trivially
 importable/runnable on its own from a plain PyTango environment - it needs
@@ -11,9 +11,11 @@ this reason; `ophyd_async.testing.float_array_value`/`int_array_value` are the
 source of truth other transports' test devices should still prefer where they
 don't have this constraint).
 
-All of these devices are served together, at fixed, predictable names, by
+Both of these are served together, at fixed, predictable names, by
 `_tango_device_servers.py` - see that module for how to actually run them up.
-Three groups, in the order they appear below:
+The `ophyd_async.tango.demo` tutorial's device servers live in
+`ophyd_async.tango.demo._tango`/`_tango_device_servers` instead - see those
+modules' docstrings for why they're kept separate from these.
 
 - `OneOfEverythingTangoDevice`: the datatype-coverage matrix (paired with the
   declarative `TangoTestDevice`).
@@ -21,17 +23,8 @@ Three groups, in the order they appear below:
   events, read-only/write-only, error paths, ...), used directly with
   `tango_signal_r`/`tango_signal_rw`/`tango_signal_w` rather than a declarative
   Device.
-- `DemoMotorDevice`/`DemoPointDetectorChannelDevice`/`DemoMultiChannelDetectorDevice`:
-  backs the `ophyd_async.tango.demo` tutorial Devices (`DemoMotor`, `DemoStage`,
-  `DemoPointDetector`, `DemoPointDetectorChannel`). Lives here, not in
-  `ophyd_async.tango.demo`, so the one fixed topology `_tango_device_servers.py`
-  serves is usable both by the tutorial and by structural tests
-  (`tests/system_tests_tango/test_base_device.py`) - there's no good reason for
-  those to instantiate different topologies of the same devices.
 """
 
-import asyncio
-import math
 import time
 import types
 from collections.abc import Sequence
@@ -41,9 +34,8 @@ from typing import Any, Generic, TypeVar
 
 import numpy as np
 import tango
-from tango import AttrDataFormat, AttrWriteType, CmdArgType, DevState, GreenMode
-from tango.asyncio import DeviceProxy
-from tango.server import Device, attribute, command, device_property
+from tango import AttrDataFormat, AttrWriteType, CmdArgType, DevState
+from tango.server import Device, attribute, command
 
 from ophyd_async.core import Array1D, DTypeScalar_co
 
@@ -516,211 +508,3 @@ class TestDevice(Device):
     )
     def enum_cmd(self, value: TestEnum) -> TestEnum:
         return value
-
-
-# --------------------------------------------------------------------------
-# Demo devices: back the ophyd_async.tango.demo tutorial Devices. See the
-# module docstring for why these live here rather than in tango.demo.
-# --------------------------------------------------------------------------
-
-
-class DemoMotorDevice(Device):
-    """Demo tango moving device."""
-
-    green_mode = GreenMode.Asyncio
-    _position = 0.0
-    _setpoint = 0.0
-    _velocity = 1.0
-    _acceleration = 1.0
-    _stop = False
-    DEVICE_CLASS_INITIAL_STATE = DevState.ON
-
-    @attribute(dtype=float, access=AttrWriteType.READ, format="%6.3f")
-    async def readback(self):
-        return self._position
-
-    @attribute(dtype=float, access=AttrWriteType.READ_WRITE, format="%6.3f")
-    async def setpoint(self):
-        return self._setpoint
-
-    async def write_setpoint(self, new_position):
-        self.set_state(DevState.MOVING)
-        self._setpoint = new_position
-        asyncio.create_task(self.move())
-
-    @attribute(dtype=float, access=AttrWriteType.READ_WRITE)
-    async def velocity(self):
-        return self._velocity
-
-    async def write_velocity(self, value: float):
-        self._velocity = value
-
-    @attribute(dtype=DevState, access=AttrWriteType.READ)
-    async def state(self):
-        return self.get_state()
-
-    @command
-    async def stop(self):
-        self._stop = True
-
-    @command
-    async def move(self):
-        self.set_state(DevState.MOVING)
-        self._stop = False
-        step = 0.1
-        while True:
-            if self._stop:
-                self._stop = False
-                break
-            if abs(self._position - self._setpoint) < abs(self._velocity * step):
-                self._position = self._setpoint
-                break
-            if self._position < self._setpoint:
-                self._position = self._position + self._velocity * step
-            else:
-                self._position = self._position - self._velocity * step
-            await asyncio.sleep(step)
-        self.set_state(DevState.ON)
-
-
-class Mode(IntEnum):
-    LOW = 0
-    HIGH = 1
-
-
-class DemoMultiChannelDetectorDevice(Device):
-    """Demo tango counting device."""
-
-    channels = device_property(dtype=int, default_value=0)
-
-    green_mode = GreenMode.Asyncio
-    _acquire_time = 0.1
-    _acquiring = False
-    _elapsed = 0.0
-
-    async def init_device(self):
-        await super().init_device()
-        self._locators = []
-        self._dps = []
-
-    @attribute(dtype=(str,), max_dim_x=32, access=AttrWriteType.READ_WRITE)
-    async def locators(self):
-        return self._locators
-
-    async def write_locators(self, value: (str)):
-        self._locators = value
-
-    @attribute(dtype=float, access=AttrWriteType.READ_WRITE)
-    async def acquire_time(self):
-        return self._acquire_time
-
-    async def write_acquire_time(self, value: float):
-        self._acquire_time = value
-
-    @attribute(dtype=bool, access=AttrWriteType.READ)
-    async def acquiring(self):
-        return self._acquiring
-
-    @attribute(dtype=float, access=AttrWriteType.READ)
-    async def elapsed(self):
-        return self._elapsed
-
-    @attribute(dtype=DevState, access=AttrWriteType.READ)
-    async def state(self):
-        return self.get_state()
-
-    @command
-    async def connect_devices(self):
-        for locator in self._locators:
-            # Connect by tango device proxy to the X motor
-            self._dps.append(await DeviceProxy(locator))  # type: ignore
-
-    @command
-    async def start(self):
-        await self._acquisition()
-
-    @command
-    async def reset(self):
-        self._elapsed = 0.0
-
-    async def _acquisition(self):
-        self._acquiring = True
-        self._elapsed = 0.0
-        step = 0.1
-        while self._elapsed < self._acquire_time:
-            self._elapsed += step
-            # Send the elapsed update to the channels
-            for dps in self._dps:
-                dps.elapsed = self._elapsed
-            await asyncio.sleep(step)
-        self._elapsed = self._acquire_time
-        for dps in self._dps:
-            dps.elapsed = self._acquire_time
-        await asyncio.sleep(step)
-        self._acquiring = False
-
-
-class DemoPointDetectorChannelDevice(Device):
-    """Demo tango counting device."""
-
-    channel: device_property = device_property(dtype=int, default_value=0)
-
-    green_mode = GreenMode.Asyncio
-    _value = 0
-    _locator_x = ""
-    _locator_y = ""
-    _elapsed = 0.0
-    _dp_x: Device | None = None
-    _dp_y = None
-    _mode: Mode = Mode.LOW
-    _energy_modes = [10, 100]
-
-    @attribute(dtype=str, access=AttrWriteType.READ_WRITE)
-    async def locator_x(self):
-        return self._locator_x
-
-    async def write_locator_x(self, value: str):
-        self._locator_x = value
-
-    @attribute(dtype=str, access=AttrWriteType.READ_WRITE)
-    async def locator_y(self):
-        return self._locator_y
-
-    async def write_locator_y(self, value: str):
-        self._locator_y = value
-
-    @attribute(dtype=Mode, access=AttrWriteType.READ_WRITE)
-    async def mode(self):
-        return self._mode
-
-    async def write_mode(self, value: Mode):
-        self._mode = value
-
-    @attribute(dtype=float, access=AttrWriteType.READ_WRITE)
-    async def elapsed(self):
-        return self._elapsed
-
-    async def write_elapsed(self, value: float):
-        self._elapsed = value
-        x: float = await self._dp_x.readback  # type: ignore
-        y: float = await self._dp_y.readback  # type: ignore
-        self._value = math.floor(
-            (
-                math.sin(x) ** self.channel  # type: ignore
-                + math.cos(x * y + self._energy_modes[self._mode])
-                + 2
-            )
-            * 2500
-            * self._elapsed
-        )  # type: ignore
-
-    @command
-    async def connect_devices(self):
-        # Connect by tango device proxy to the X motor
-        self._dp_x = await DeviceProxy(self._locator_x)  # type: ignore
-        # Connect by tango device proxy to the Y motor
-        self._dp_y = await DeviceProxy(self._locator_y)  # type: ignore
-
-    @attribute(dtype=int, access=AttrWriteType.READ)
-    async def value(self):
-        return self._value
